🎨 update: Refactor tutorial type import
Refactor to import `Tutorial` type from `types/tutorial`. Added a new
type definition file to centralize tutorial type definitions.
