+++ src/pages/admin/tutorials/EditTutorial.tsx
@@ -4,7 +4,7 @@ import { useAuth } from 'context/AuthContext'
 import { Editor } from '@tinymce/tinymce-react'
 import PageBanner from 'components/Common/PageBanner'
 import { Link, useNavigate, useParams } from 'react-router-dom'
-import { Tutorial } from 'pages/Tutorials'
+import { Tutorial } from 'types/tutorial'
 import { toast } from 'react-toastify'
 import { ToastDefaultOptions } from 'utils/toastOptions'
 import { slugify } from 'utils/slugify'
