+++ src/pages/admin/tutorials/ListTutorial.tsx
@@ -11,8 +11,8 @@ import { useAuth } from 'context/AuthContext'
 import { Link } from 'react-router-dom'
 import { FaEye, FaPencil } from 'react-icons/fa6'
 import PageBanner from 'components/Common/PageBanner'
-import { Tutorial } from 'pages/tutorials/List'
 import { slugify } from 'utils/slugify'
+import { Tutorial } from 'types/tutorial'
 
 
 const ListTutorials: React.FC = () => {
