+++ src/pages/tutorials/List.tsx
@@ -10,25 +10,7 @@ import { slugify } from 'utils/slugify'
 import defaultTutoImg from 'assets/images/defaultTutoImg.png'
 import { stripHtmlTags } from 'utils/stripHtmlTags'
 import ImageLoader from 'components/Common/ImageLoader'
-
-export type Tutorial = {
-  id: number;
-  userId: number;
-  categoryId: string;
-  title: string;
-  image?: string;
-  content: string;
-  views: number;
-  upvote: number;
-  slug: string;
-  createdAt: Date;
-  commentCount: number;
-  validated: boolean;
-  user?: {
-    nickname: string;
-    avatar: string;
-  }
-};
+import { Tutorial } from 'types/tutorial'
 
 const TutorialsList: React.FC = () => {
   const [tutorials, setTutorials] = useState<Tutorial[]>([])
